made edit buttons viewable on mobile and madea reusable component for expandable text
