+++ frontend/src/components/ui/display/index.js
@@ -12,3 +12,4 @@ export { default as Divider } from './Divider';
 export { default as StatItem } from './StatItem';
 export { default as SectionHeader } from './SectionHeader';
 export { default as LinkifyText } from './LinkifyText';
+export { default as ExpandableText } from './ExpandableText';
